fix: made SoundEvent directly accessible
